chore: seperated the main.rs to smaller modules
